Add axe metatype
Axe metatype:
Has a break option that allows an item to be unbreakable at 0 durability,
but it also becomes useless.

Has a onHitTarget effects list that are potion effects added
to the target when hitting an entity.

Has a onHitUser effects list that are potion effects added to the user
when hitting an entity

Has a wear effects list that are potion effects added to the item holder
every tick while holding the item